Refactor interactive tests
