+++ gradle.properties
@@ -11,3 +11,4 @@ junitJupiterVersion=5.9.2
 junitJupiterParamsVersion=5.9.2
 junitJupiterEngineVersion=5.9.2
 testfxVersion=4.0.16-alpha
+testfxMonocleVersion=jdk-12.0.1+2
