+++ src/test/kotlin/solve/interactive/scene/view/landmarks/InteractiveTestsSetup.kt
@@ -1,0 +1,19 @@
+package solve.interactive.scene.view.landmarks
+
+import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.extension.ExtendWith
+import org.testfx.framework.junit5.ApplicationExtension
+
+@ExtendWith(ApplicationExtension::class)
+internal class InteractiveTestsSetup {
+    companion object {
+        @JvmStatic
+        @BeforeAll
+        fun setUpAll() {
+            System.setProperty("testfx.robot", "glass")
+            System.setProperty("testfx.headless", "true")
+            System.setProperty("prism.order", "sw")
+            System.setProperty("prism.text", "t2k")
+        }
+    }
+}
